Fix NPE loading Personal Stats for Pokemon with no sprite resource
This has been a really long-standing bug, hopefully it's fixed for good!

+++ pk3DS/Misc/Patch.cs
@@ -109,7 +109,7 @@ namespace pk3DS
             catch (Exception ex)
             { 
                 Util.Error("Could not create patch:", ex.ToString());
-                Directory.Delete(patchFolder, true);
+                if (Directory.Exists(patchFolder)) Directory.Delete(patchFolder, true);
             }
         }
